# Dyno: cache function signature instantiations, reduce cache impact of some queries