--- a/frontend/lib/resolution/scope-queries.cpp
+++ b/frontend/lib/resolution/scope-queries.cpp
@@ -3238,11 +3238,25 @@ struct GatherMentionedModules {
   const Module* inMod = nullptr;
   std::set<const Scope*> scopes;
   std::set<ID> idSet;
-  std::vector<ID> idVec;
+  std::vector<ID>idVec;
+  std::vector<std::pair<const Scope*, const AstNode*>> scopeStack;
 
   GatherMentionedModules(Context* context, const Module* inMod)
     : context(context), inMod(inMod)
   { }
+
+  void enterScope(const AstNode* node);
+  void exitScope(const AstNode* node);
+
+  const Scope* currentScope() {
+    CHPL_ASSERT(!scopeStack.empty());
+    if (scopeStack.back().first == nullptr) {
+      scopeStack.back().first =
+        scopeForId(context, scopeStack.back().second->id());
+    }
+
+    return scopeStack.back().first;
+  }
 
   void gatherModuleId(const ID& id);
   const Scope* lookupAndGather(const Scope* scope, const AstNode* ast,
@@ -3273,16 +3287,35 @@ struct GatherMentionedModules {
 
   // do not delve into submodules
   bool enter(const Module* m) {
+    enterScope(m);
     return (m == inMod); // visit the requested module only
   }
-  void exit(const Module* m) { }
+  void exit(const Module* m) {
+    exitScope(m);
+  }
 
   // traverse through anything else
   bool enter(const AstNode* ast) {
+    enterScope(ast);
     return true;
   }
-  void exit(const AstNode* ast) { }
+  void exit(const AstNode* ast) {
+    exitScope(ast);
+  }
 };
+
+void GatherMentionedModules::enterScope(const AstNode* node) {
+  if (resolution::createsScope(node->tag())) {
+    scopeStack.emplace_back(nullptr, node);
+  }
+}
+
+void GatherMentionedModules::exitScope(const AstNode* node) {
+  if (resolution::createsScope(node->tag())) {
+    CHPL_ASSERT(!scopeStack.empty());
+    scopeStack.pop_back();
+  }
+}
 
 // save the module used/imported to idSet / idVec
 void GatherMentionedModules::gatherModuleId(const ID& id) {
@@ -3433,12 +3466,12 @@ void GatherMentionedModules::processDot(const Dot* d) {
   // TODO: ideally this would use the result of scope resolution.
   // It does not yet because running the scope resolver in all
   // these cases would currently lead to compilation slowdowns
-  const Scope* scope = scopeForId(context, d->id());
+  const Scope* scope = currentScope();
   gatherAndFindScope(scope, d);
 }
 
 void GatherMentionedModules::processUseImport(const AstNode* ast) {
-  const Scope* scope = scopeForId(context, ast->id());
+  const Scope* scope = currentScope();
   if (scope && scope->containsUseImport()) {
     auto p = scopes.insert(scope);
     if (p.second) {